fixing DOMMatrix error
